moved diana into uist, preliminary stuff on fortuna
