+++ crates/rotala/src/orderbook/mod.rs
@@ -1,4 +1,3 @@
 //! Internal order execution logic that is used by exchanges. Orderbooks have their own data
 //! formats that exchanges have to implement.
-pub mod diana;
 pub mod fortuna;
